refactor: ロビー作成ページをpagesに移動
