+++ src/pages/lobby-creation/api/create-lobby.ts
@@ -1,7 +1,7 @@
 import type { SupabaseClient } from "@supabase/supabase-js";
-import { convertDBLobbyToLobby } from "../lib/convert-db-lobby-to-lobby";
-import { creationLobySchema } from "../schemas";
-import type { Lobby, LobbyCreationDate } from "../types";
+import { convertDBLobbyToLobby } from "../../../features/lobby/lib/convert-db-lobby-to-lobby";
+import { creationLobySchema } from "../../../features/lobby/schemas";
+import type { Lobby, LobbyCreationDate } from "../../../features/lobby/types";
 
 export async function createLobby(
 	supabase: SupabaseClient,
